lettuce: init: Don't mount /system in charge mode
This fixes reboot when holding the power button in charge mode.  When
/system is mounted, ueventd spawns a child that spins forever waiting
for the modem firmware to appear.  This in turn causes the reboot to
fail.  But when /system is not mounted, the ueventd child doesn't seem
to get that far and terminates, so reboot works.

It appears the only reason to mount /system is for /system/bin/vm_bms,
which is not present on this device.

Change-Id: Ieaf3a4b9d78a4f31f7cc7d8c085fda6952899ef8
(cherry picked from commit c935baddbcda724a886bf41475f90d53d0382861)